% Generated by roxygen2 (4.0.2): do not edit by hand
\name{WRSEL}
\alias{WRSEL}
\title{Weighted Rank Squared Error loss}
\usage{
WRSEL(X, num_nodes = NULL)

WRSEL(X, num_nodes = NULL)
}
\arguments{
\item{X}{samples from the posterior, size \code{n} by \code{N}}

\item{num_nodes}{the number of nodes to use on multi-core architecture (needs parallel package)}

\item{weights}{the weight function, must be a vector of length 1 or length \code{n}}

\item{X}{samples from the posterior, size \code{n} by \code{N}}

\item{weights}{the weight function, must be a vector of length 1 or length \code{n}}

\item{num_nodes}{the number of nodes to use on multi-core architecture (needs parallel package)}
}
\value{
function
}
\description{
Given a set of samples of an n-dimensional random variable, \code{WRSEL} returns a function which, when supplied with a weight vector, returns
optimal estimates under a rank-weighted squared error loss, see references for details.
}
\details{
Requires the Matrix package. Returns a function with which we can apply any weight
}
\examples{
### Simulate a three variable system, where the three variables are Gaussian with sd = 1 and means c(-1,0,1)
n <- 3
N <- 100000
X <- matrix(rnorm(n*N),n,N)
X <- X + c(-1,0,1)
### Create the WRSEL function which we can then supply with weight vectors to obtain the optimal estimates
WRSEL_fun <- WRSEL(X = X,num_nodes = NULL)
### Create a weight vector
weights <- c(1,0,0)
### Find and print the optimal estimates
lambda_hat <- WRSEL_fun(weights)
print(lambda_hat)
}
\references{
Wright, Deanne L., Hal S. Stern, and Noel Cressie. "Loss functions for estimation of extrema with an application to disease mapping." Canadian Journal of Statistics 31.3 (2003): 251-266.
}
\keyword{error}
\keyword{exceedances}
\keyword{loss,}
\keyword{rank}
\keyword{squared}
\keyword{weighted}

